+++ src/PageRouter.js
@@ -9,217 +9,6 @@ import Nav from "./lib/Nav";
 import HomePage from "./pages/HomePage";
 import ProfilePage from "./pages/ProfilePage";
 
-const sampleUser = {
-  email: "johsn@smith.com",
-  name: "John Smith",
-  journal: {
-    weeks: [
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-    ],
-  },
-  notes: [
-    {
-      header: "header",
-      content: "content",
-    },
-    {
-      header: "header",
-      content: "content",
-    },
-  ],
-};
-
 const links = [
   { title: "Home", href: "/" },
   { title: "Profile", href: "/profile" },
