+++ src/components/journal/JournalDash.js
@@ -1,0 +1,14 @@
+import React, { Component } from "react";
+
+class JournalDash extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return <div>Journal Dashboard...</div>;
+  }
+}
+
+export default JournalDash;
